Updates to .sspace file generation, optimisations to vector creation and bug fixes for sparse hash vectors
M    src/edu/ucla/sspace/mains/GenericMain.java

- Updated so that the output may either be a file or a directory.  This allows
  all *Main classes to save their output to a specific file, which makes keeping
  track of batch sspace results much easier

M    src/edu/ucla/sspace/purandare/PurandareFirstOrder.java

- Fixed bug where matrix was being unnecessarily written to disk

- Improved logging slightly

- TODO: fix the ClutoClustering code which is coming in from the reflective-ri
  branch

M    src/edu/ucla/sspace/text/WordIterator.java

- Fix to close the BufferedReader when the iterator has no more tokens

M    src/edu/ucla/sspace/common/StaticSemanticSpace.java

- Updated readSparseBinary to take advantage of the new CompactSparseVector
  constructor

M    src/edu/ucla/sspace/vector/VectorMath.java

- Fixed javadoc

M    src/edu/ucla/sspace/vector/SparseHashVector.java

- Fixed bug where double values in the range (-1,1) were being treated as 0
  values

M    src/edu/ucla/sspace/vector/CompactSparseVector.java

- Added overloaded constructor to allow creating a vector from its non-zero
  indices and values.

M    src/edu/ucla/sspace/util/SparseHashArray.java

- updated getElementIndices to return the indices in sorted order.  Once the
  method is called, the indices are cached until the array is modified again

- Fixed bug in set() null values were not correctly accounted for in the element
  indices

M    src/edu/ucla/sspace/util/SparseIntArray.java
M    src/edu/ucla/sspace/util/SparseDoubleArray.java

- Added overloaded constructor to allow creating an array from its non-zero
  indices and values.


M    build.xml

- Added unit tests

A    test/edu/ucla/sspace/text/FilteredIteratorTests.java

- Added tests for the FilteredIterator class

+++ sspace/src/edu/ucla/sspace/vector/SparseHashVector.java
@@ -90,6 +90,6 @@ public class SparseHashVector<T extends Number>
      * {@inheritDoc}
      */
     public void set(int index, Number value) {
-        vector.set(index, (value.intValue() == 0) ? null : value);
+        vector.set(index, (value.doubleValue() == 0d) ? null : value);
     }
 }
